+++ lib/actionizer/configuration_error.rb
@@ -1,4 +1,0 @@
-module Actionizer
-  class ConfigurationError < StandardError
-  end
-end