+++ lib/irie/configuration_error.rb
@@ -1,0 +1,4 @@
+module Irie
+  class ConfigurationError < StandardError
+  end
+end