+++ test/dummy/config/routes.rb
@@ -1,5 +1,13 @@
 Dummy::Application.routes.draw do
   resources :barfoos
   resources :foobars
-  resources :special_barfoos
+
+  namespace :example do
+    namespace :company do
+      #scope '/magic' do
+        resources :foobars
+      #end
+      resources :special_barfoos
+    end
+  end
 end